robust improvements to the lin_to_lin solver.  wrote a wrapper around the mptype0,2 and mptype1 methods, and now correctly adjust the precision of the randomizer matrix, linears, patches, etc, for AMP.
now also resolves the case in lin_to_lin where you have 0 linears to move to, and it returns an empty but properly initialized witness_set.

the double and mp data for the witness set are contained in the same object.  witness_set_d *should* be renamed witness_set.

next step is to merge in Wenrui's code, and make another commit.  

then, to make robust linprod_to_detjac