restructure handles; separate space and color; performance audit
this blob of changes enables the colorfn to use, or not, the projected space data or raw space data when using a custom colorfn in conjunction with a custom coordinate projection.  i also took the opportunity to run the plotting code through the profiler, and made many performance improvements. the code is significantly faster as a result.  most of the performance things were, embarassingly enough, stringifying indices to make the labels in the plot.  however, there were indeed some problems where the data being used was not always what was most correct.  hence, i modified the way the data is preprocessed before plotting.