using HOMVAR=1 only in select places
after using it in full decomposition for dihedral was extremely slow.  i think that the real patch is only needed when doing real homotopies

it needs to be instrumented in the surface connect the dots routine, still, because there there are multiple variable groups, possibly with several patches, and the structure must be preserved.  to come tomorrow...