added option to keep all vertices when converting to mesh
this is towards working with data in grasshopper, as I need the curve and surface pieces to agree on the vertices.  otherwise I'd have to play some very stupid games trying to figure out which vertex is which, and there's enough of that bullshit in the cellular decomposition algorithm as it is, lol.