added input file for skew cylinders
this input file describes a system which Bertini_real computes incorrectly because it's missing a step.  essentially, this demonstrates an algorithmic flaw.

we know how to fix it.  it's to compute the intersection points downstairs, in projection space, in addition to the critical points of the critical curve.

silviana: see the photos from the board work on August 20, 2024