decomposition saves pathnumbers
this was a huge set of changes.  many file affected

- now vertices in .vertex files also record the path numbers that ended there.
- also, now bertini_real (but not sampler) have the option `-savepaths`, which will save the paths to the `./paths` folder

I pushed metadata holding into every point holder.  but it's still a bit wrong, in that the vertex set used for this decomposition holds points that are computed many times.  so holding just one metadata per point is incorrect.

but at least i get the path numbers.

issues:
* some path numbers are repeated.  this is a bug
* this commit does NOT include parsing of the path nubmers into Python or Matlab
* the metadata is lost between beritni_real and sampler.    the path numbers stay, but the per-point metadata is lost.  i regard this as ok, because we should really save the metadata with EVERY encounter, not per-point.
* i have not tested parallel computation.  something in send/receive could be fucked.
* the paths folder is not emptied between runs, so old data could remain.
* probably a lot more.  this is WIP.

---

i also made some adjustments to const-correctness and const& while i was mucking around.

i've really come to see the many ways in  which this code is weak.  i wish I could start over.  alas, there is a lot of momentum, and starting over would be impossible.  the only way i'm starting over is with bertini2 as the core.  there's so much trash code in here due to b1's use of pointers everywhere.